+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Simplify the page creation wizard by reducing possibilities and automating object relations
   based on the navigation context:
+  * Automatically link courses with an organization when created from this organization's page,
   * Create nested categories automatically when visiting the parent category page,
   * Show button to create course runs only when visiting a course page and automatically link
     the course run to the current course,